+++ EssentialFeediOSTests/FeedViewControllerTests.swift
@@ -426,7 +426,7 @@ private extension UIButton {
 }
 
 private extension UIRefreshControl {
-    func simulatePullToRefresh() {
+    func  simulatePullToRefresh() {
         allTargets.forEach({ target in
             actions(forTarget: target, forControlEvent: .valueChanged)?.forEach {
                 (target as NSObject).perform(Selector($0))
@@ -435,15 +435,4 @@ private extension UIRefreshControl {
     }
 }
 
-private extension UIImage {
-    static func make(withColor color: UIColor) -> UIImage {
-        let rect = CGRect(x: 0, y: 0, width: 1, height: 1)
-        UIGraphicsBeginImageContext(rect.size)
-        let context = UIGraphicsGetCurrentContext()!
-        context.setFillColor(color.cgColor)
-        context.fill(rect)
-        let img = UIGraphicsGetImageFromCurrentImageContext()
-        UIGraphicsEndImageContext()
-        return img!
-    }
-}
+
